+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Change Log
  *  Feature: Add automatic Gzip wrapping and `isGzipped()` method for all assets
  *  Improvement: Add Loader classes for better adapter usage
  *  Feature: Add `asUri()` method to assets to get path as a Uri String
+ *  Improvement: Update cache handling and add max size settings to the builder
 
 ## Version 0.3